+++ chapter_deep-learning-basics/logistic_regression.py
@@ -3,23 +3,7 @@
 # @Author  : douwentao
 import numpy as np
 import tensorflow as tf
-
-
-def obt_shape(x):
-    # 组合静态shape和动态shape
-    shape_list = []
-    static_shape_list = x.get_shape().as_list()
-    none_index_list = []
-    for i, val in enumerate(static_shape_list):
-        if val == None:
-            none_index_list.append(i)
-    if len(none_index_list) == 0:
-        return static_shape_list
-    dynamic_shape_list = tf.shape(x)
-    for i in range(len(static_shape_list)):
-        if i in none_index_list:
-            shape_list.append(dynamic_shape_list[i])
-    return shape_list
+from utils import obt_shape
 
 
 def sigmoid(x):
@@ -80,6 +64,7 @@ def logistic_regression_test():
         w = w - tf.multiply(learning_rate, update_w)
         b = b - tf.multiply(learning_rate, update_b)
     sess.close()
+
 
 if __name__ == '__main__':
     logistic_regression_test()
