+++ chapter_deep-learning-basics/softmax_regression.py
@@ -3,25 +3,7 @@
 # @Author  : douwentao
 import numpy as np
 import tensorflow as tf
-
-
-def obt_shape(x):
-    # 组合静态shape和动态shape
-    shape_list = []
-    static_shape_list = x.get_shape().as_list()
-    none_index_list = []
-    for i, val in enumerate(static_shape_list):
-        if val == None:
-            none_index_list.append(i)
-    if len(none_index_list) == 0:
-        return static_shape_list
-    dynamic_shape_list = tf.shape(x)
-    for i in range(len(static_shape_list)):
-        if i in none_index_list:
-            shape_list.append(dynamic_shape_list[i])
-        else:
-            shape_list.append(static_shape_list[i])
-    return shape_list
+from utils import obt_shape
 
 
 def prepare_data(feature_dimension):
@@ -104,6 +86,7 @@ def softmax_regression():
             update_w = tf.add(cur_update_w, update_w)
         w = w - tf.multiply(learning_rate, update_w)
     sess.close()
+
 
 if __name__ == '__main__':
     softmax_regression()
